fiveq_test/requirements.txt
Django==1.5.1
psycopg2==2.5
django-taggit==0.10a1
